// web/src/api/generated/models/workflow-template-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 工作流模板响应
 */
export interface WorkflowTemplateResponse {
    /**
     * 模板ID
     */
    'id'?: number;
    /**
     * 模板名称
     */
    'name'?: string;
    /**
     * 模板描述
     */
    'description'?: string;
    /**
     * 分类
     */
    'category'?: string;
    /**
     * 图标URL
     */
    'icon'?: string;
    /**
     * 工作流定义JSON
     */
    'definition'?: string;
    /**
     * 标签
     */
    'tags'?: Array<string>;
    /**
     * 创建者ID
     */
    'creatorId'?: number;
    /**
     * 使用次数
     */
    'usageCount'?: number;
    /**
     * 创建时间
     */
    'createTime'?: string;
    'public'?: boolean;
    'system'?: boolean;
}